+++ .github/codecov.yml
@@ -1,5 +1,9 @@
 coverage:
   status:
+    project:
+      default:
+        target: auto
+        threshold: 1%
     patch:
       default:
         target: 90%
